feat(DTFS2-7298): small renamings
