+++ libs/common/src/constants/tfm/deal-cancellation-status.ts
