+++ libs/common/src/constants/tfm/index.ts
@@ -1,5 +1,5 @@
 export * from './team-ids';
 export * from './teams';
 export * from './deal-stage';
-export * from './tfm-deal-cancellation-status';
+export * from './deal-cancellation-status';
 export * from './entra-id-user-groups';
