+++ libs/common/src/types/tfm/deal-cancellation-status.ts
