+++ libs/common/src/types/tfm/index.ts
@@ -1,5 +1,5 @@
 export * from './deal-submission-type';
 export * from './team-id';
 export * from './team';
-export * from './tfm-deal-cancellation-status';
+export * from './deal-cancellation-status';
 export * from './entra-id-user';
